+++ frontend/src/scenes/session-recordings/filters/SavedFiltersStates.tsx
@@ -4,10 +4,10 @@ import { LemonBanner } from 'lib/lemon-ui/LemonBanner'
 
 import { ReplayTabs } from '~/types'
 
-import { savedSessionRecordingPlaylistsLogic } from '../saved-playlists/savedSessionRecordingPlaylistsLogic'
+import { sessionRecordingSavedFiltersLogic } from '../filters/sessionRecordingSavedFiltersLogic'
 
 export function SavedFiltersEmptyState(): JSX.Element {
-    const playlistsLogic = savedSessionRecordingPlaylistsLogic({ tab: ReplayTabs.Home })
+    const playlistsLogic = sessionRecordingSavedFiltersLogic({ tab: ReplayTabs.Home })
     const { loadPlaylistsFailed } = useValues(playlistsLogic)
     return loadPlaylistsFailed ? (
         <LemonBanner type="error">Error while trying to load saved filters.</LemonBanner>
@@ -24,7 +24,7 @@ export function SavedFiltersEmptyState(): JSX.Element {
 }
 
 export function SavedFiltersLoadingState(): JSX.Element {
-    const playlistsLogic = savedSessionRecordingPlaylistsLogic({ tab: ReplayTabs.Home })
+    const playlistsLogic = sessionRecordingSavedFiltersLogic({ tab: ReplayTabs.Home })
     const { loadPlaylistsFailed } = useValues(playlistsLogic)
     return loadPlaylistsFailed ? (
         <LemonBanner type="error">Error while trying to load saved filters.</LemonBanner>
