+++ frontend/src/scenes/session-recordings/playlist/SessionRecordingsPlaylistSettings.tsx
@@ -5,11 +5,10 @@ import { LemonBadge, LemonButton, LemonCheckbox, LemonInput, LemonModal, Spinner
 
 import { getAccessControlDisabledReason } from 'lib/components/AccessControlAction'
 import { LemonMenuItem } from 'lib/lemon-ui/LemonMenu/LemonMenu'
+import { sessionRecordingCollectionsLogic } from 'scenes/session-recordings/collections/sessionRecordingCollectionsLogic'
 import { SettingsBar, SettingsMenu } from 'scenes/session-recordings/components/PanelSettings'
-import { savedSessionRecordingPlaylistsLogic } from 'scenes/session-recordings/saved-playlists/savedSessionRecordingPlaylistsLogic'
 
 import { AccessControlLevel, AccessControlResourceType, RecordingUniversalFilters } from '~/types'
-import { ReplayTabs } from '~/types'
 
 import { playerSettingsLogic } from '../player/playerSettingsLogic'
 import {
@@ -187,7 +186,7 @@ function NewCollectionModal(): JSX.Element {
         useValues(sessionRecordingsPlaylistLogic)
     const { setIsNewCollectionDialogOpen, setNewCollectionName, handleCreateNewCollectionBulkAdd } =
         useActions(sessionRecordingsPlaylistLogic)
-    const { loadPlaylists } = useActions(savedSessionRecordingPlaylistsLogic({ tab: ReplayTabs.Playlists }))
+    const { loadPlaylists } = useActions(sessionRecordingCollectionsLogic)
 
     const handleClose = (): void => {
         setIsNewCollectionDialogOpen(false)
@@ -243,9 +242,7 @@ export function SessionRecordingsPlaylistTopSettings({
 }): JSX.Element {
     const { autoplayDirection } = useValues(playerSettingsLogic)
     const { setAutoplayDirection } = useActions(playerSettingsLogic)
-    const { playlists, playlistsLoading } = useValues(
-        savedSessionRecordingPlaylistsLogic({ tab: ReplayTabs.Playlists })
-    )
+    const { playlists, playlistsLoading } = useValues(sessionRecordingCollectionsLogic)
     const { selectedRecordingsIds, sessionRecordings, pinnedRecordings } = useValues(sessionRecordingsPlaylistLogic)
     const {
         handleBulkAddToPlaylist,
